Revert "Feature/basic file retrieval from git"
