+++ lib/specifications/getFeatureFile.js
@@ -3,7 +3,7 @@
 var fs = require("q-io/fs"); // https://github.com/kriskowal/q-io
 
 /**
- * Get teh contents of a feature file from disk.
+ * Get teh contents of a feature file.
  *
  * @return a promise for the file contents.
  */
