+++ views/feature-files.hbs
@@ -1,0 +1,33 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+
+    <title>Your Specifications</title>
+    <meta name="description" content="A service for exposing Gherkin feature files as specifications.">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <link rel="stylesheet" href="public/css/modified-normalize.css">
+    <link rel="stylesheet" href="public/css/alphabeta.css">
+    <link rel="stylesheet" href="public/css/main.css">
+
+    <!-- Old IE hack for main elements-->
+    <script>document.createElement('main');</script>
+  </head>
+  <body>
+    <div class="alphabeta"></div>
+    <main>
+      <header>
+        <h1>Available Specifications</h1>
+      </header>
+      <section>
+        {{#each paths}}
+          <p><a class="speclink" href="{{this.featurePath}}">{{this.featureName}}</a></p>
+        {{else}}
+          <p>No specifications found</p>
+        {{/each}}
+      </section>
+    </main>
+  </body>
+</html>
