--- views/feature.hbs
@@ -1,14 +1,0 @@
-{{> layout_start}}
-<main>
-  <header>
-    <h1>Your Feature</h1>
-  </header>
-  <section>
-    {{#each featureLines}}
-      {{this}}<br>
-    {{else}}
-      <p>Sorry, no feature found</p>
-    {{/each}}
-  </section>
-</main>
-{{> layout_end}}
