--- views/features.hbs
@@ -1,14 +1,0 @@
-{{> layout_start}}
-<main>
-  <header>
-    <h1>Available Specifications</h1>
-  </header>
-  <section>
-    {{#each paths}}
-      <p><a class="speclink" href="{{this.featurePath}}">{{this.featureName}}</a></p>
-    {{else}}
-      <p>No specifications found</p>
-    {{/each}}
-  </section>
-</main>
-{{> layout_end}}
