--- views/index.hbs
@@ -1,15 +1,0 @@
-{{> layout_start}}
-<main class="project-management group">
-  <h1>Your Projects</h1>
-  <section>
-    <h2>Get features from a project</h2>
-    <form action="/" method="get">
-      <label>Project Git repo URL (https only please): <input type="text" name="repo_url" /></label>
-      <button class="call-to-action">Submit</button>
-    </form>
-  </section>
-  <section>
-    <h2><a href="/features">Your features</a></h2>
-  </section>
-</main>
-{{> layout_end}}
