--- views/partials/layout-end.hbs
@@ -1,2 +1,0 @@
-</body>
-</html>
